feat: remove types.ts, remove valid adr check
